requirements_dev.txt
notebook
py-youtube
ensure
mkdocs-material
ipywidgets
tqdm
python-dotenv
pip-chill

# dev requirements -
pytest==7.1.3
tox==3.25.1
black==22.8.0
flake8==5.0.4
mypy==0.971
-e .